% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marathon.R
\name{get_marathon}
\alias{get_marathon}
\title{Scrape marathon data}
\usage{
get_marathon(event = "marathon", gender = "women", year = 2018)
}
\arguments{
\item{event}{The event (marathon or half marathon) to scrape. Defaults to marathon.}

\item{gender}{The gender (men or women) to scrape. Defaults to women.}

\item{year}{Year to scrape. (Data only available from 2001).}
}
\description{
This function allows you to scrape marathon and half marathon data from the IAAF website.
}
\examples{
get_marathon(event = "half-marathon", gender = "men", year = 2012)

get_marathon(event = "marathon", gender = "women", year = 2018)
}
\keyword{track}
